Candle option is now long (not int)
Candle-histo has no no line when there is no content

line width is set as well

updated fre tuts

+++ graf2d/graf/inc/TCandle.h
@@ -33,7 +33,7 @@ const Int_t kNMAXPOINTS = 2000;  // Max outliers per candle
 class TCandle : public TAttLine, public TAttFill, public TAttMarker {
 public:
    //Candle Option
-   enum CandleOption : int {
+   enum CandleOption : long {
       kNoOption = 0,
       kBox = 1,
       kMedianLine = 10,
